refactor: 공지사항 메인&상세 모델 파일 병합
